Corrected Statistics Problem for LC Members (refs #902).
The Query was designed to look up a users National Group. However, some LC members are not a member of the National Group, but should still be allowed to work.
Correction meant that another Query was added which found the National Group belonging to the same Member country as the Local Committee.