Added support for changing monitoring (refs #100).
Monitoring is one of the internal features that was added early on, but never exposed via the API. The functionality has now been added, and can be accessed. Further, an error was found in the contact listing, where it read suspended information out.
Note; This update contains a DB update as well!