Extended GroupType changes to database.
Adding some of the flags to the database, so it can be better controlled, if a Group may have a public and/or private mailing list. Policy is simple - if someone is "on" a public or private mailinglist and the Group may have one, then it'll work, otherwise not. The concept of aliases is linked with the Private Groups, which can only belong to a single user anyway.