Added FethInternationalGroups.
For the Committee's, it must be possible to select all the International Groups, with their current owner.
Note, that the fetch Operations are currently very simple and will *only* fetch those Groups which have an Owner. As it is a prerequisite that Groups always have an owner, this should suffice. If some Committees or International Groups exists without an Owner, it is considered a data error and must be fixed manually in the database.