Corrected AccountName change test.
The test was failing, as it assumed that members may change their own names. This is now allowed, as it was earlier misused to hand over accounts and thus caused a data protection problem - since some data was not properly deleted before the handover. Generally, members should learn to add accounts for all users and then use these properly.