Added Private/Public Mailing list Flags to Groups.
For the future IWS Views to be used by the mail server, the Groups must have a flag so it is possible to tell if the Group is having a public and/or private mailing list. Which is again checked against the Group Type.
The testing is not yet completed, so the feature is currently considered experimental.