Added new Offer & Exchange Types.
After a lengthy discussion with various Committees, it was a wish to have the Type of Offer (Limited, Reserved or Other) as well as a type of Exchange (AC, IntraWeb, COBE) added. So it is possible to change either the Offer Type or Exchange Type, without affecting the Reference Number.