Added Offer & Exchange Types to the Persistence layer (refs #930).
Note, this means it is possible to create and update offers with the new Offer & Exchange Type. However, currently, the Offer & Exchange Types are preset in the Offer Object, which they shouldn't be. This was left so initially, to ensure that existing test will work with existing test data - until these have been corrected.
The updates for the database is almost in place, what is missing is the correction of existing Reference Numbers, which causes problems due to conflicts in the numbers given!
New rule to no longer silently correct the reference number is not yet applied as it will cause massive test errors, which also needs to be corrected.