Corrections to API to ensure error handling is present in WSDL.
The previous error object was not present in the generated WSDL file, and hence also not present in any generatet source code. Since the error handling is vital to ensure that the client is getting the current state from IWS, it has to be present.
Fix was to add an XML Type, which made the Abstract parts look weird, so to ensure that the code is still working properly, the AbstractFallible Class was dropped and the FallibleResponse is used instead as the default super Object.
Note; Although this is a rather large API change, the API itself remains untouched.