Updates CSV with OfferType (refs #930).
Offers can now be uploaded and downloaded with the OfferType. ExchangeType is omitted, since it is purely for internal registration and rules, and should just be part of the sharing of offers.
Additionally, the old silent Update rule for the Offer Reference Number (see Trac #372) has been reverted, following a request from Olga Legacka, current. Now, if the Reference Number is invalid, an error is returned.