Added Folder Stubs (refs #946).
Added methods fetchFolder and processFolder, with their required request & response Objects. Also added a first version of the Folder Object, which holds both a list of Folders and a list of Files. And finally, added a privacy flag to the File Object, so it is easier to control who may see a file.