Prepared corrections for the CSV error handling (refs #966).
Note, the changes have not yet been tested and are thus also not in production. The new variant is using a different way of dealing with errors, which is hopefully cleaner and easier to understand.
It should be noted, that the other corrections are not damaging anything, but they are not proven to have resolved the issue, since they are untested and we are still awaiting that Korea will hopefully give us their CSV, so we can properly test that we can replicate the error and that the fix is working.