Archive Migration is working (refs #946).
From the 10.580 IW3 Files, 10.289 IWS Folders & Files were persisted, and 291 was skipped, since it is not possible to ascertain ownership anymore. Complete migration was done in 2½ minutes.
Problem: how to deal with privacy - this still remain unanswered. By default, all files are Private - but Folders are public, only the owners can view private files in public folders, and only group members can view protedtec files in a public folder.
Incomplete; The fetching of Files has not been modified yet, so it is not yet possible to view a File, however - that is a minor detail.