Adding proof-of-concept code to interceptor.
The interceptor can be used for more than just profiling, we can also use them for updating Sessions. It can even be extended with more things like data for error handling later on - such as full request and response information. However, it must be clarified how this works with data protection. Still it can be an invaluable source for generating anonymized data for actual error cases.