Added corrections to DB model & logic (refs #946).
With this update, the database model should be complete for the Library. However, the functionality has not been completed due to complexities with the queries. Which means that for now, a reduced functionality is presented and later, once we have some better Folder/File Views in place in the database, we can present the full scope.
For the moment the listing of files is deliberately disabled, since the scope rules must be completely in place first - as we may otherwise be presenting offers that was not meant for the public!