Updates to sessions and logging (refs #977).
It was annoying to see two log records for each request, one from the profiler and one from the method. By removing the profiler and make a better generic log generator - it means that the log messages are now better.
Additionally, the logic for updating sessions has been improved to use the DAO method instead of doing the JQL directly. This means that the logic is better synchronized.
Missing is still to improve the handling of active sessions in the Session Monitor. So Sessions will automatically be considered dead if unused for a period of n minutes.