Restructuring of Folder logic (refs #946).
The logic for the Folders and files was slowly growing out of control. It is restructured to better allowe control over DAO parts so they can be moved to a new DAO class later. Further, the logic was a bit reviewed to verify that the structure is logical and makes sense.
  Class still needs a refactoring since some of the methods have exceeded a reasonable and readable length.
  All open issues is marked with a TODO.
  Note; This is still alpha quality. No testing has been done!