Added proof-of-concept for WS support in the Client (refs #939).
To properly support WebServices in the IWS Client, the WS files must be automatically generated from the WSDL files, and not just a single, but several. This commit should add the proof-of-concept things for it. It uses CXF, and the two WSDL files is generated via WildFly.
It should be noted, that this client hasn't been tested against the test or production servers, but that is hopefully a minor issue.