Removed Role "Guest".
The Guest Role was used as a "Member without mail privileges" in IW3, but in IWS with the mail flags - it has become irrelevant, and is just another thing to maintain needlessly, so with this, it is no longer in use. Corrections to the tables (Id's) and in the production database is still pending. But can be fixed release independent.