Internal changes to our Date & DateTime Objects (refs #939).
The Date and DateTime Objects were not properly exposed by the WebServices, and as the Objects also required a more stringent mapping with proper XML marshalling & unmarshalling - I simply replaced the internal representation.
Added more mapping to show how it works. The Client I'm building is showing how to connect, list permissions and logout. The mapping is not complete, since the Permissions requires that all of the core user Objects is mapped over. However, it is a step forward :-)