Fixed SQL Constraint Violation Exception.
The purpose of the Session Request Bean is to help improving the internal error handling, by saving enough information to properly create an (anonymized) test. To do this, we save certain things, one of them is the error message from the underlying Exception.
For an unknown reason, we had a case in the production where the error message was null - thus causing an SQL Constraint Violation Exception to be thrown. To avoid these and also ensure that we can improve the error logging, a default text is set if nothing else is present.