Bugfix related to deleting UserGroup records for Member groups.
Member Groups is a central place to control Members access to the system, and a user must belong to a Member Group to be able to use the system. The bugfix prevents users from being able to delete users from Member Groups.
Additionally, the delete boolean flag was replaced with an Action, and the Request Object was made Actionable.