Sample Client for the WS (main method) completed (refs #939).
The main method has been restructured, so it is cleaner and easier to read and understand. The Class itself is now instantiable, and contains public methods to perform each request. To use it, simple correct the username/password as well as the host & port. Then it should run cleanly against any IWS instance. It has been tested against both Glassfish3 and WildFly8.