Added try-catch blocks to all WS methods.
If an error occurred as part of the transactional logic, then the error was flying out of IWS, rather than being caught and handled. As we're moving to pure WS based communication for external access, it makes sense to add the try-catch blocks additionally here, to catch Transactional problems and log it with full stacktrace and only return a simple error, stating that something is wrong without explaining it.
Although we should not hide information - telling the details about an unknown error is generally not a good idea. Hence, the information is hidden of.
Later, the profiling parts from the EJB module should also be moved over to the WS module. But as long as IW4 is not invoking the WS, but rather the EJB's, we'll keept it as it is.