Minor bugfix related to updating username.
The log message when updating the username referred to the password! This is incorrect, so it has been corrected. At the same time a minor bugfix was added, for those users who have upper-case letters in their passwords. Since the IWS saves the hash of a lowercase version of the password.