Added support for Properties file in IW4.
Note, that with this commit, it is now possible to control how IW4 is accessing IWS using the properties file iw4.properties, which should be located in the GLASSFISH config directory. A default file is added to the IWS Accessories.
Default setting, using WebService on localhost port 8080. WildFly is configured to use port 9080, so this means that IW4 is expecting that IWS is running under Glassfish as well. However, this should change soon!