Removal of legacy settings.
Board members were not able to create users, as some old legacy information was still present in the database. Besides this, to make it easier to read the emergency list, a new view was added (not mapped to any entity, nor should it be) and finally, the update script for the production database has been reset, since the changes have been applied.