Set a default empty list for OffersIds.
There's numerous errors in IW4, regarding the list of Id's being null - although the Id's are optional, there's a null check. So the correction is simple, set the default value to an empty list, rather than null.
Also, added/updated JavaDoc and added input validation checks to the OfferCSVDownloadRequest Object.