Correction to the NC's mailing list query.
The Query was used by the mail migration script to update the NC's mailing list on the mail server. However, the Board was missing from it - not a big deal as all Board members were present in various countries so they would be found by different criterias.
However, for the International Groups, the problem was (is), that only the Owner is on the List - meaning that it is not possible to grant someone else easily access the IDT list and also to the NC's list without reassigning the Ower (Coordinator) role.
New version is a hybrid, old version had around 270 members, new one has about 300. So a few more people will get the mails - which is hopefully not going to be a big problem.