Fixed bug in resetPassword (refs #100, #1004).
The resetPassword Request uses the general Password Object, which is also used by the updatePassword Request. To avoid having to make strange checks to verify the identification information (old password or password token), the Object was altered so both Requests will use both the new Password and the identification information.
  As no tests existed for the resetPassword request, a couple of new integration tests have been added to verify that it works as expected.
  The Change is an API update, so both IWS & IW4 must be updated.