Added new Filedata Entity.
The files were originally suppose to be stored in the database, but a former administrator insisted on storing things on the disc for backup reasons. however, as we have zero backup strategies, but a fully functional database replication, the former decision is reversed.
Storing of files in the database is done in a separate table, which will hold a foreign key to the primay file data. This way, we can build our JPA queries so regardless of laziness they won't fetch the file data - unless explicitly needed.