Applied correction to the handling of e-mail addresses (refs #984).
The handling is simply extended with a better error handling, although no other mechanisms have been applied to catch it. It seems that the error occurs so seldom that adding more controls will only clutter up the complexity of the logic. If it happens more frequently, then more corrections can be made.
However, the primary goal is to never see NPE's in the IWS.