Changes in logic for Exchange Year & Offer Expiration (refs #1020, #1022).
A couple of issues with Offers were corrected. First, the problem with Offers automatically expiring on September 1st, regardless. This was addressed and resolved. Second issue was that Offers expired on the deadline, not after! Meaning that it was not possible to nominate on the last day.
Problems with the change of Exchange Year have also been addressed, mainly so the Offers from the current calendar year will not disappear on September 1st. This has been the cause of many problems for users. See discussion in Trac task #1022 for details.