Corrections to the NC's mailing list (refs #1003).
A couple of corrections to the NC's mailing list following a change to the rules. Note, that there was a problem with a JUnit test, which is failing in local Maven builds. This needs to be investigated further  - but to avoid loosing work, and because manual tests in IW4 seems to show no problems, the commit is made anyway.