Extended error handling when processing Employers (refs #1023).
The problem where an Employer is attempted to be updated, but the values match one Employer but the Id differs - has been causing problems. IWS will not support such a case, as it is clearly an error. However, it should not fail with an SQL Constraint Violation Exception.
This commit will not fix the problem itself, but only serve to make it more clear what is going wrong and why.