Removal of dependencies that is not needed.
Currently, the IWS deployment file is fairly large, mainly because Hibernate is present, but this is not needed in IWS when deployed under WildFly, since WildFly is using Hibernate - it was only needed for Glassfish deployments. However, the IWS is no longer capable of being deployed under Glassfish, since there were some compatibility problems with JPA. And also because we wish to no longer run in a Java7 environment.
Second dependency removed is JodaTime, since it was a very old version used that was deprecated. Upgrading meant rewriting, so instead I removed it and replaced it with standard Java functionality. IWS cannot yet be upgraded to Java8 in the source code level, since the API is needed by IW4, running under Glassfish3.
The last dependency to remove is Freemarker, since it is about 1MB large, but only used for simple replacements.