Extended Fetch Offer Request with more fields (refs #100, #1022).
The IWS was altered, so it could retrieve Offers from both the current year and exchange year, meaning that from September until December, Offers from both this and next year is listed. However, this caused confusion, so the Request Object has been extended with a flag to switch this off.
At the sametime, the Request Object has also been extended with a feature to filter Offers based on current State. This was a request from Czech Republic, since they found it pointless to filter the offers client side.