Fixed problem with deleting users.
There's been several errors in the IWS logs, related to deleting Users, who have Requests records associated with a Session. The Request Objects serve as a way to collect error information to help trace problems in the IWS. Such records will be analyzed and later deleted manually if the error is benign. If error is an actual problem, the classic approach is made - create test based on findings and fix bug.
Fix was simple - only delete sessions without error information.