Corrections to the WSDL Definitions.
Using an Analyser, the WSDL files have been improved, so they have no errors, warnings or hints. The only thing left is to convert the files from RPC to Document, but doing so requires some quite substantial changes, so that will be left for later. For more information about the RPC vs Document style, see this StackOverflow answer: http://stackoverflow.com/a/28322548