Temporary correction to Exchange Logic.
IAESTE Switzerland reported a problem when trying to force expiration of an Offer, by setting the Nomination Deadline to a date in the past. This is a User action that the existing logic did not handle, so new functionality has been added to support it.
It should be noted, that this fix is temporary - it helps the Users, but should not be allowed. A proper verification of the Exchange Logic is needed.