Removed the AddressEntity from the PersonEntity merge (refs #1028).
The IWS have some features for merging changes, thus making it easier to update entities. However, the merge methods should never refer to other Entities, since their state cannot be verified as part of the merge. Under a special circumstance, this triggered a bug since we already updated the Address, and have set it into the Person, and then we merged a different Person entity into it. This meant that we left an orphan entity dangling, and Hibernate went crazy with errors.