Correcting the WSDL Host + Port in the WildFly configuration (refs #1027).
With this, the WSDL will always be correctly generated according to the viewing host, so if someone has it deployed on localhost and is accessing it via port 9080 (insecure), then the address is localhost:9080. However in production, when accessed via iws.iaeste.net:9443 (standard), then the address is reading the same.