Corrected weird error when reading a folder (refs #946).
When reading folders, the debugger showed that sub-folders & files were found. Problem was just that they weren't returned, although they were set in the Response Object! I've changed the handling of the lists in the Response, so there's always a list and it is filled by copying over the content from the other list. This way, the test was working and we get a list of Folders & Files back.