Added new response Object for the processCommittee request (refs #929).
For many requests to the processing of Committees, it is good to also have the information back regarding both the Committee & NS. Additionally, the Group DTO have been extended with the new MailReply information for both Public & Private Lists and additionally a ParentId.
Note, most of the changes was made to the API, not to the implementation!