Correction to lookup of users to suspend (refs #1050).
The list of users to be suspended included those who have never logged in, although their accounts have been activated. This meant that new accounts which have never been used is automatically suspended! The fix is to divide the list of accounts into 2, the "regular" accounts, i.e. accounts where the user have already logged in, and accounts which don't have a login.
Only those users who have not used their otherwise active account within the configured amount of days will be added to the first list of accounts to be suspended. The rest will be left.