Added missing validation check (refs #1048).
The missing validation check meant that when trying to create or update an sharing group, then an NPE was thrown if no groups were present. Since the request is considered invalid if no groups are present, it was simple to fix. However, IW4 also needs updating to make sure it handles the errors better!