Fixed a bug when retrieving a list of countries.
The old logic was using the View for fetching Countries, which linked in the Group System, to find the NS. However, this meant that for any non-member country to the list was empty. So far, this little bug went unnoticed, as our testing only included a list of member countries, so our test data simply didn't cover this scenario.
The update is using the entity instead, when we're using the Country Type rather than the Committee Type.