Corrections to Offer State when sharing (refs #1054).
The problem in the IntraWeb, when it comes to sharing is that the entire logic is flawed and the process bad. Meaning that it is pending a rewrite! However, the problem reported dealt with the specific case when someone changed part of the sharing information, then the new flag was set based on country information - which is incorrect. A new logic was added to verify that Expired offers with deadline in the future is changed to shared and offers with deadline in the past is set to expired.