Removed support for MD5 Checksums.
The support for MD5 checksums was needed to allow users to access the IW with their IW3 passwords. However, we're not 1+ year after the launch of IWS/IW4, so either accounts have been updated or suspended. So there is no longer any need to support this highly insecure standard anymore.