Initial version of the Leargas Client.
Please note, that the work on this Client began in the Spring but was halted as Leargas cannot or will not use a Java component to fill their database. So the Client has never been completed.
The Source code can hopefully serve as inspiration to others on how to integrate a local system with the Production IWS.
Note, the code is being committed, but without any guarantees that anything in it will work, if it burns down the house, you're warned!