Added static alias table, altered Users & Offers (refs #1058, #1050, #1060).
In preparation for merging Committees, where we need to support static aliases for a period. And also because we currently have some added, which is not controlled anywhere. A new table has been added for these.
Since the EULA is bound to change, a new flag for the lates approved EULA version has been added, note that no logic is currently planned for supporting this flag.
A thing that was missing was also a timestamp for the last time the password was changed. With this, we can later on add a rule for enforcing password changes.