Corrections to CSV error handling (refs #1077).
The CSV error handling was not the best, as it caused loads of stack traces in the log files, and as each stachtrace requires analysis - it is stealing time. The Error handling have now been improved, both when it comes to dealing with enumerated types, but also if the Header is invalid.