Extended StateBean with a Session Reset flag.
The handling of sessions internally in the IWS has been a topic for discussion for a long time. Either, they sessions are too long or too short. And it was also a discussion if we should reset sessions upon restart or not, as it is a good feature in development, but terrible in production.
Now, a flag has been added to the iws.properties file and propagated inside the IWS. Via this flag, it is possible to control this setting. And it also means that it is possible to redeploy IWS without affecting users who are currently logged in.