Pruning the FitNesse module.
The purpose of the FitNesse module is to provide a set of pages, which can both test and document the behavior of the IWS. However, due to too many other things, and a lack of proper testers, this is now considered pruned, i.e. pending a decision if it should be discontinued completely or if the time can be found to complete it.
  For all, who wish to integrate their national or local systems with the IWS, this module will be a considerable help and source of information, as it can quickly be extended to display different behavior so new Clients can verify their behavior against this.