Added Reflective Standard Methods, to reduce boilerplate code.
Most of the equals, hashCode and toString methods were just generated using the default generator in the IDE. This meant that there was loads of code cluttering up the DTO's especially, which made them more unreadable. Changes in these were often forgotten - and this meant that some fields were not present. Additionally, the format of the toString value was inconsistent.
Using an exiting reflective framework for these, like Apache Commons, is not a good idea, since we have fields which needs to be protected. There's fields, which should be present on hashCode and equals but not in toString or vice versa. Instead, a simple Reflective engine has been created, which will handle the generation, and with a new Annotation, it is possible to fine-grain the control so it fits the needs we have in IWS.