Almost completed migration to new Reflective Standard method engine.
Almost all classes and tests have been correctly updated. Note, that enumerated types may cause test errors as the hashCode generator for an enumerated type is not constant based, but based on the JVM instance.
Further, one Class have not been migrated yet, the Role DTO. Reason is because there seems to be a flaw in the Reflective equals method, which causes test problems with tests invoking the fetchRoles request.
However, as 99% of all code seems to work perfectly with the new engine, the migration is being committed - and the Role will for now remain the old version, until the flaw can be identified and closed.