First part of the mailing list changes (refs #1082).
With this commit, the new tables are in place. the old tables removed - we're no longer referring to the mailing list DB, so we also cannot update it. What is lacking, is updating the internal logic, so we can create new records, and add the changes to the State Bean so it is possible to run updates frequently.