Added newlines to mails (refs #1089).
When dropping the Freemarker Library, the code was re-written to just use a StreamReader, unfortunately, when doing this - I forgot to add the newlines as well.
For some strange reason, the testing when the Freemarker Library was revoked showed that things worked. However, with this check, it should hopefully be better.