Updated constants from Interface to Class.
We originally prefered using Interfaces for the constants, as there were no intention to actually implement it, and it meant that the constants were purely constants with no logic. However, to avoid that someone in the future makes the mistake and implements the constants, rather than imports them - they are converted.