Added comment regarding AC properties (refs #933).
Adding the logic to check for the Exchange Type rules, is fairly simple. But, it affects testing in a negative way, since there will be periods where te builds will fail. A set of new tests must be written, and existing must be extended with a way to allow adjustable settings.