Fixed problem with reading templates.
The error with using NIO seems obvious in retrospect. NIO relies on file path's and can thus not work with files in a JAR! Instead, the ancient Scanner trick is used. Strange that nobody have managed to introduce a better reading of resources than a trick which is over a decade old!
JUnit tests of this all run through clean, and testing with sending using a local WildFly configured for sending also shows that it works.