Added listname to private groups.
Earlier, Private Groups were only used to avoid mixing UserId's and GroupId's, like we had in IW3. However, so far they have not been used for mailing lists, since Users were only having an Alias. However, with the changes to the mail setup, the Private Groups is used to create private aliases from - so now the Private Group also requires the listname.
Note, that the list name in Groups must be without the domain, since this is handled via the mailing list generator.