Initial commit with the new mailing list and subscription rules in place.
The work with the mailing list is finally seeing an and. There's two showstoppers left (subscribe/unsubscribe and mayWrite for AnnounceList), otherwise only documentation and massive testing is missing. However, the new setup is already an improvement over the old system, since the synchronization is done via the StateBean. So regardless of changes, the mail tables should always be correct.