Added test and applied a minor fix (refs #1100).
Note, that the fix applied was that the returned Object when updating an Offer was not the saved Object, but the given Object, so it didn't reflect what was actually in the Database. Additionally to test, the Offer was also retrieved via the FetchOffer method, but here the list of Identifiers was ignored, so this also had to be fixed.
So, two minor fixes were made, but the cause was not found - from the look, it is puzzling why the State can be changed, and the only cause may be if the deployment of the latest release was not completely wiping the old code.