Added NS Check to State Bean before Suspending Accounts.
The State Bean was so far suspending accounts if they were inactive for longer than n days (configurable via the properties file). However, this had the undesirable side-effect that some National Secretaries were also suspended, although they shouldn't be. As a consequence, some countries lost access to the IntraWeb.
This commit adds a check to see if a person is the owner of either an Administrative, Member, National or International Group, if so - we're logging this and ignoring the Account.