Removed FitNesse permanently.
FitNesse is a good test tool for backend systems. Which the IWS is. It provides mechanisms to invoke a service via self-implemented fixtures, and the result is thus a combined test and documentation.
Problem is, that the FitNesse module was never seriously maintained, fixtures were added, but after a while the module was in state Pruned, meaning that it was left as it was without maintenance. Bringing it back to life is therefore a major undertaking.
It will make perfect sense to have FitNesse in the future, but given the state of the current source, it will be easier to start from Scratch than trying to get this to work - also because only a proof-of-concept test existed, nothing else!