Correction to Mailing List Write flag.
The MayWrite flag was added as the Board wished to have a way to control who could write to a private list. Although the flag is correctly working in the IWS, it was only set when creating a record. With this fix, it is also being updated when the flag changes for a User.