Moved Instance Password Salt into the properties file.
It was an old mistake to have the IWS Instance Password Salt hardcoded in the logic, it meant that it could not be changed. The Heartbleed attack showed that it is good if all Passwords can be instantly void, and with the salt in the Properties file, it can be accomplished.