Relicensed IWS to APL 2.0.
The developers who have a stake in the current version of the IWS, have all agreed to move the IWS to the Apache License, Version 2.0 (APL 2.0). Thus making the IWS an Open Source Project.
Please note, that it is not being relicensed as an Open Source Project in the hope that it will attract more developers - it is done to open up the process of development for others and make it more democratic.