Added new Test Class to generate Bot data.
Many Committees have requested that a Portal is made available, where Students can read out redacted shared Offers for a given Country, for this to work, a new account is created via this test for each committee - called "Committee Offer Bot". The Account starts with status "SUSPENDED", meaning that it is an opt-in option for Committees to decide if they wish to have this feature or not, and they can ay any time change it themselves.